Move functions to resolve circular import
