+++ InfrastructureManager/agents/openstack_agent.py
@@ -7,6 +7,8 @@ import os
 from urlparse import urlparse
 from utils import utils
 
+from appscale.common.constants import KEY_DIRECTORY
+
 __author__ = 'dario nascimento'
 __email__ = 'dario.nascimento@tecnico.ulisboa.pt'
 
@@ -43,7 +45,7 @@ class OpenStackAgent(EC2Agent):
     keyname = parameters[self.PARAM_KEYNAME]
     group = parameters[self.PARAM_GROUP]
 
-    key_path = '{}/{}.key'.format(utils.KEY_DIRECTORY, keyname)
+    key_path = '{}/{}.key'.format(KEY_DIRECTORY, keyname)
     ssh_key = os.path.abspath(key_path)
     utils.log('About to spawn OpenStack instances - ' \
               'Expecting to find a key at {0}'.format(ssh_key))
